Use dataclass fields in InferenceRPUConfig (#37)
Revise `InferenceRPUConfig` so the fields are modeled as attributes in
the same way as the rest of `RPUConfigs`. The difference is that those
fields are not present in the constructor (via the `init=False flag`):
it is not meant to make them fully read-only, but at least provide
some assurances about the right field usage while reducing some
complexity and keeping the configs a bit more homogeneous.